requirements.txt
numpy
pandas
boto3
scikit-learn
mysql-connector-python
python-dotenv
joblib